Moved screens to components directory
