+++ little-guy-creator-app/components/createScreen.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { StyleSheet, TouchableOpacity, Text, View } from "react-native";
 
-import {styles} from './Styles.js';
+import {styles} from '../styless.js';
 
 const CreateScreen = ({ navigation, route }) => {
     return (
