+++ little-guy-creator-app/components/editScreen.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { StyleSheet, TouchableOpacity, Text, View } from "react-native";
 
-import {styles} from './Styles.js';
+import {styles} from '../styless.js';
 
 const EditScreen = ({ navigation, route }) => {
     return (
